update import paths, adjust types
