+++ src/module/vue/bondset-sheet.vue
@@ -39,20 +39,17 @@
 </template>
 
 <script setup lang="ts">
-import { computed, inject, provide, reactive, watch, ref } from 'vue'
+import { computed, inject, provide, reactive } from 'vue'
 import { $ItemKey, ItemKey } from './provisions'
-import type { BondsetDataPropertiesData } from '../item/itemtypes'
 import MceEditor from './components/mce-editor.vue'
 import IronBtn from './components/buttons/iron-btn.vue'
 
-const props = defineProps<{ data: { item: any } }>()
+const props = defineProps<{ data: { item: ItemSource<'bondset'> } }>()
 provide(ItemKey, computed(() => props.data.item) as any)
 
 const $item = inject($ItemKey)
 
-const bonds = computed(
-	() => (props.data.item.system as BondsetDataPropertiesData).bonds
-)
+const bonds = computed(() => props.data.item.system.bonds)
 
 const data = reactive({
 	selectedBondIndex: -1,
@@ -72,7 +69,7 @@ function selectBondIndex(i: number) {
 }
 
 async function deleteBond(i) {
-	const system = props.data.item.system as BondsetDataPropertiesData
+	const system = props.data.item.system
 	const bonds = Object.values(system.bonds)
 	bonds.splice(i, 1)
 	await $item?.update({ system: { bonds } })
@@ -85,7 +82,7 @@ async function deleteBond(i) {
 }
 
 async function addBond() {
-	const system = props.data.item.system as BondsetDataPropertiesData
+	const system = props.data.item.system
 	const bonds = Object.values(system.bonds)
 	bonds.push({ name: '', notes: '' })
 	await $item?.update({ system: { bonds } })
