+++ src/module/vue/shared-sheet.vue
@@ -31,7 +31,6 @@ import { $ActorKey, ActorKey } from './provisions'
 import Bonds from './components/bonds.vue'
 import MceEditor from './components/mce-editor.vue'
 import ActiveCompletedProgresses from 'component:progress/active-completed-progresses.vue'
-import type { BondsetDataPropertiesData } from '../item/itemtypes'
 import SheetBasic from './sheet-basic.vue'
 import ConditionMeter from './components/resource-meter/condition-meter.vue'
 import { IronswornSettings } from '../helpers/settings.js'
@@ -43,10 +42,10 @@ provide(ActorKey, computed(() => props.data.actor) as any)
 const $actor = inject($ActorKey)
 
 const hasBonds = computed(() => {
-	const bonds = (
-		props.data.actor.items.find((x) => x.type === 'bondset') as any
-	)?.system as BondsetDataPropertiesData | undefined
-	const markedBonds = bonds?.bonds?.length
+	const bonds = props.data.actor.items.find((x) => x.type === 'bondset') as
+		| ItemSource<'bondset'>
+		| undefined
+	const markedBonds = bonds?.system.bonds?.length
 	return markedBonds && markedBonds > 0
 })
